[Autodiscovery] Rename FileSystem to AutodiscoveryFinder, to make class unique
