+++ packages/autodiscovery/config/config.yaml
@@ -8,7 +8,7 @@ services:
         exclude:
             - '../src/Contract/'
             - '../src/HttpKernel/AutodiscoveryKernel.php'
-            - '../src/FileSystem.php'
+            - '../src/Finder/*'
             - '../src/*/*Autodiscoverer.php'
             - '../src/Discovery.php'
